Split AdminSettingsStore into focused modules (656 → 141 lines)
The store had four distinct concerns tangled in one file: env
probes, password helpers, the load-time settings merge with three
multi-account migrations, user CRUD with auth, and secret masking.
Each now lives in its own file under services/admin-settings/:

- env.ts             production-required env vars + probes
- auth-helpers.ts    hashPassword / verifyPassword / defaultAuthSettings
                     / createDefaultAdminUser / sanitizeUser /
                     normalizeUsers
- mergeSettings.ts   load-time defaults + legacy → multi-account
                     migration for github / email / google; the three
                     migrators are now named helpers (mergeGitHub,
                     mergeEmail, mergeGoogle) instead of inline IIFEs
- io.ts              loadAdminSettings + updateAdminSettings + the
                     internal writeSettings primitive
- userCrud.ts        listManagedUsers / createManagedUser /
                     updateManagedUser / findAdminUser /
                     authenticateManagedUser /
                     updateCurrentUserCredentials / updateAuthSettings
- publicMasking.ts   getPublicAdminSettings — strips per-integration
                     and per-custom-field secrets before sending to
                     the admin UI

AdminSettingsStore.ts keeps the small update wrappers
(updateAppSettings, resetAppSettings, updatePersonalizationSettings,
updateIntegrationSettings) and re-exports everything else so the 10+
existing import sites don't need to change. All files under 220 lines.
